Docs now reference methods via the public ArgumentParser class instead of internal mixin classes
