+++ jsonargparse/_instantiation.py
@@ -1,4 +1,5 @@
 import inspect
+from typing import Union
 
 from ._common import (
     ClassType,
@@ -7,11 +8,114 @@ from ._common import (
     applied_instantiation_links,
     class_instantiators,
     is_subclass,
+    parser_context,
 )
+from ._namespace import Namespace, get_value_and_parent, split_key
 
 __all__ = ["add_instantiator"]
 
 _global_class_instantiators: InstantiatorsDictType = {}
+
+
+class InstantiateMethod:
+    def instantiate(
+        self,
+        cfg: Namespace,
+        instantiate_groups: bool = True,
+    ) -> Namespace:
+        """Instantiates all signature components in a configuration namespace.
+
+        Processes the configuration recursively, converting each signature
+        component registered with the parser into its corresponding Python
+        object:
+
+        - **Class/subclass type arguments** (``add_argument`` with a class type
+          or ``add_class_arguments``/``add_subclass_arguments``): An object with
+          ``class_path`` and optionally ``init_args`` is replaced by an instance
+          of the referenced class, created by calling
+          ``class_type(**init_args)``. For the case of classes with disabled
+          subclasses, the namespace can have directly the init args without the
+          ``class_path`` + ``init_args`` wrapper.
+
+        - **Callable type arguments**: A dot-import string pointing to a
+          function or method is resolved to the callable object. When
+          ``class_path``/``init_args`` is given instead and the class
+          instantiates into a callable (or is a subclass of the callable's
+          return type), the result is either a class instance or — when not all
+          call arguments are provided yet — a :func:`functools.partial` bound to
+          the given ``init_args``.
+
+        - **Instantiation order**: Components are processed in the order
+          determined by argument links applied on instantiation.
+
+        Args:
+            cfg: The configuration object to use. Must have been produced by
+                one of the ``parse_*`` methods and not modified in a way that
+                breaks the structure expected by the parser.
+            instantiate_groups: Whether class groups should be instantiated.
+
+        Returns:
+            A new configuration object where every registered signature
+            component has been replaced by its corresponding Python object.
+        """
+        from ._actions import _ActionConfigLoad, filter_non_parsing_actions
+        from ._core import ArgumentGroup
+        from ._link_arguments import ActionLink
+        from ._subcommands import get_subcommand
+        from ._typehints import ActionTypeHint
+
+        components: list[Union[ActionTypeHint, _ActionConfigLoad, ArgumentGroup]] = []
+        for action in filter_non_parsing_actions(self._actions):  # type: ignore[attr-defined]
+            if isinstance(action, ActionTypeHint):
+                components.append(action)
+            elif isinstance(action, ActionLink) and isinstance(action.target[1], ActionTypeHint):
+                components.append(action.target[1])
+
+        if instantiate_groups:
+            skip = {c.dest for c in components}
+            groups = [
+                g
+                for g in self._action_groups  # type: ignore[attr-defined]
+                if hasattr(g, "instantiate_class") and g.dest not in skip
+            ]
+            components.extend(groups)
+
+        components.sort(key=lambda x: -len(split_key(x.dest)))  # type: ignore[arg-type]
+        order = ActionLink.instantiation_order(self)
+        components = ActionLink.reorder(order, components)
+
+        cfg = cfg.clone(with_meta=False)
+        for component in components:
+            ActionLink.apply_instantiation_links(self, cfg, target=component.dest)
+            if isinstance(component, ActionTypeHint):
+                try:
+                    value, parent, key = get_value_and_parent(cfg, component.dest)
+                except (KeyError, AttributeError):
+                    pass
+                else:
+                    if value is not None:
+                        with parser_context(
+                            parent_parser=self,
+                            nested_links=ActionLink.get_nested_links(self, component),
+                            class_instantiators=get_class_instantiators(self),
+                            applied_instantiation_links=cfg.get("__applied_instantiation_links__"),
+                        ):
+                            parent[key] = component.instantiate_classes(value)
+            else:
+                with parser_context(
+                    load_value_mode=self.parser_mode,  # type: ignore[attr-defined]
+                    class_instantiators=get_class_instantiators(self),
+                    applied_instantiation_links=cfg.get("__applied_instantiation_links__"),
+                ):
+                    component.instantiate_class(component, cfg)
+
+        ActionLink.apply_instantiation_links(self, cfg, order=order)
+
+        subcommand, subparser = get_subcommand(self, cfg, fail_no_subcommand=False)  # type: ignore[arg-type]
+        if subcommand is not None and subparser is not None:
+            cfg[subcommand] = subparser.instantiate(cfg[subcommand], instantiate_groups=instantiate_groups)
+
+        return cfg
 
 
 def add_instantiator(
