+++ jsonargparse/_signatures.py
@@ -32,9 +32,6 @@ from ._typehints import (
 from ._util import NoneType, get_import_path, get_private_kwargs, get_typehint_origin, iter_to_set_str
 from .typing import _LazyInitBaseClass, register_pydantic_type
 
-__all__ = ["SignatureArguments"]
-
-
 kinds = inspect._ParameterKind
 inspect_empty = inspect._empty
